// src/app/@shared/@enums/player-class.enum.ts
export enum PlayerClass {
    Bard = "Bard",
    Beastlord = "Beastlord",
    Berserker = "Berserker",
    Cleric = "Cleric",
    Druid = "Druid",
    Enchanter = "Enchanter",
    Magician = "Magician",
    Monk = "Monk",
    Necromancer = "Necromancer",
    Paladin = "Paladin",
    Ranger = "Ranger",
    Rogue = "Rogue",
    Shadowknight = "Shadowknight",
    Shaman = "Shaman",
    Warrior = "Warrior",
    Wizard = "Wizard",
    Unknown = "Unknown"
}